requirement.txt
pandas==1.5.3
paramiko==3.4.0
biopython==1.83
